Purge expired certificates and secrets
Checks daily for any expired certificates and secrets on the managed application registration and purges them.

+++ Kerbee/Graph/ApplicationService.cs
@@ -237,6 +237,12 @@ internal class ApplicationService : IApplicationService
 
     public async Task RemoveKeyAsync(Application application)
     {
+        if (application.KeyId is null)
+        {
+            _logger.LogWarning("Application {applicationId} does not have a key", application.Id);
+            return;
+        }
+
         if (application.KeyType == KeyType.Certificate)
         {
             await _graphService.RemoveCertificateAsync(application.Id.ToString(), new Guid(application.KeyId));
@@ -247,5 +253,31 @@ internal class ApplicationService : IApplicationService
             await _graphService.RemoveSecretAsync(application.Id.ToString(), new Guid(application.KeyId));
             await _secretClient.StartDeleteSecretAsync(application.KeyName);
         }
+    }
+
+    public async Task PurgeKeys(Application application)
+    {
+        var graphApplication = await _graphService.GetApplicationAsync(application.Id.ToString());
+
+        if (graphApplication is null)
+        {
+            return;
+        }
+
+        // Remove expired secrets
+        var expiredSecrets = graphApplication.PasswordCredentials?
+            .Where(x => x.EndDateTime <= DateTime.UtcNow)
+            .Where(x => x.KeyId.HasValue)
+            .ToArray() ?? Array.Empty<Microsoft.Graph.Models.PasswordCredential>();
+
+        await Task.WhenAll(expiredSecrets.Select(x => _graphService.RemoveSecretAsync(application.Id.ToString(), x.KeyId!.Value)));
+
+        // Remove expired certificates
+        var expiredCertificates = graphApplication.KeyCredentials?
+            .Where(x => x.EndDateTime <= DateTime.UtcNow)
+            .Where(x => x.KeyId.HasValue)
+            .ToArray() ?? Array.Empty<Microsoft.Graph.Models.KeyCredential>();
+
+        await Task.WhenAll(expiredCertificates.Select(x => _graphService.RemoveCertificateAsync(application.Id.ToString(), x.KeyId!.Value)));
     }
 }
